+++ client/App.js
@@ -25,12 +25,8 @@ export default function App() {
   return (
     <RecoilRoot>
       <NavigationContainer>
-        <Stack.Navigator initialRouteName="AccountBook">
-          <Stack.Screen
-            name="Main"
-            component={Main}
-            options={{ headerShown: false }}
-          />
+        <Stack.Navigator initialRouteName="TransactionHistory">
+          <Stack.Screen name="Main" component={Main} options={{ headerShown: false }} />
 
           <Stack.Screen name="AccountBook" component={AccountBook} />
           <Stack.Screen
@@ -108,10 +104,7 @@ export default function App() {
           />
 
           <Stack.Screen name="SignUp" component={SignUp} />
-          <Stack.Screen
-            name="TransactionHistory"
-            component={TransactionHistory}
-          />
+          <Stack.Screen name="TransactionHistory" component={TransactionHistory} />
         </Stack.Navigator>
       </NavigationContainer>
     </RecoilRoot>
